requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==1.0.0
python-dotenv==0.19.0
pyyaml==6.0
streamlit>=1.32.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14 